Move ProviderLabel and TargetIcon to common
