+++ assets/js/common/ProviderLabel/ProviderLabel.jsx
