+++ assets/js/common/ProviderLabel/ProviderLabel.stories.jsx
