+++ assets/js/common/ProviderLabel/index.js
