+++ assets/js/common/TargetIcon/TargetIcon.jsx
