+++ assets/js/common/TargetIcon/TargetIcon.stories.jsx
