+++ assets/js/common/TargetIcon/TargetIcon.test.jsx
