+++ assets/js/common/TargetIcon/index.js
