+++ assets/js/pages/ChecksCatalog/ChecksCatalog.jsx
@@ -13,8 +13,8 @@ import { clusterTypes, getClusterTypeLabel } from '@lib/model/clusters';
 import Accordion from '@common/Accordion';
 import PageHeader from '@common/PageHeader';
 import Select, { createOptionRenderer, OPTION_ALL } from '@components/Select';
-import ProviderLabel from '@components/ProviderLabel';
-import TargetIcon from '@components/TargetIcon';
+import ProviderLabel from '@common/ProviderLabel';
+import TargetIcon from '@common/TargetIcon';
 import CatalogContainer from './CatalogContainer';
 import CheckItem from './CheckItem';
