+++ assets/js/pages/ExecutionResults/CheckResultDetail/CheckResultInfoBox.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import ListView from '@common/ListView';
-import ProviderLabel from '@components/ProviderLabel';
+import ProviderLabel from '@common/ProviderLabel';
 import { TARGET_CLUSTER, TARGET_HOST } from '@lib/model';
 
 const targetTypeString = {
